refactor: RecruitingManagementPage 컴포넌트화
